python_demo/pyproject.toml
[tool.poetry]
name = "grpc-python"
version = "0.1.0"
description = ""
authors = ["jacantwell <jasper66018@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.67.0"
grpcio-tools = "^1.67.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
